refactor: make log/event structure clear
Centralizing the log/event structure in the event objects in `types.ts`
to make it completely clear what the structure of these logs/events are,
and to simplify the emission of events in `loggers.ts`.

While doing this refactoring it also became clear that
logLoopDetectionDisabled was only emitting an event for clearcut
logging, not for open telemetry as well. Following up with @SandyTao520
about #8503 where this was added, it doesn't seem there was a strong
reason for this, so this was updated to be consistent also.

Fixes #10347